EFRS-890: Create endpoints for face detection - created 'sdk' package for working with external services - created and used FacesApiClient/FacesRestClient to work with Faces service to decouple from Feign - refactored code
