+++ api/src/main/java/com/exadel/frs/core/trainservice/component/migration/MigrationComponent.java
@@ -20,8 +20,8 @@ import com.exadel.frs.core.trainservice.entity.Face.Embedding;
 import com.exadel.frs.core.trainservice.entity.Image;
 import com.exadel.frs.core.trainservice.repository.FacesRepository;
 import com.exadel.frs.core.trainservice.repository.ImagesRepository;
-import com.exadel.frs.core.trainservice.system.feign.FeignClientFactory;
-import com.exadel.frs.core.trainservice.system.feign.faces.FacesServiceClient;
+import com.exadel.frs.core.trainservice.sdk.config.FeignClientFactory;
+import com.exadel.frs.core.trainservice.sdk.faces.feign.FacesFeignClient;
 import com.exadel.frs.core.trainservice.util.MultipartFileData;
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
@@ -59,7 +59,7 @@ public class MigrationComponent {
     }
 
     private void processFaces(final String url) {
-        val migrationServerFeignClient = feignClientFactory.getFeignClient(FacesServiceClient.class, url);
+        val migrationServerFeignClient = feignClientFactory.getFeignClient(FacesFeignClient.class, url);
         val migrationCalculatorVersion = migrationServerFeignClient.getStatus().getCalculatorVersion();
 
         log.info("Calculating embedding for faces");
@@ -70,21 +70,20 @@ public class MigrationComponent {
                 val image = imagesRepository.findById(face.getId()).orElse(new Image());
                 val file = new MultipartFileData(image.getRawImg(), face.getFaceName(), null);
 
+                val faceEmbedding = new Embedding();
                 try {
                     val scanResponse = migrationServerFeignClient.scanFaces(file, 1, null);
-
                     val embeddings = scanResponse.getResult().stream()
                                                  .findFirst().orElseThrow()
                                                  .getEmbedding();
-                    face.setEmbedding(new Embedding(embeddings, scanResponse.getCalculatorVersion()));
-
-                    facesRepository.save(face);
+                    faceEmbedding.setEmbeddings(embeddings);
+                    faceEmbedding.setCalculatorVersion(scanResponse.getCalculatorVersion());
                 } catch (FeignException.InternalServerError | FeignException.BadRequest error) {
-                    log.error("{} during processing facename {} with id {}", error.toString(), face.getFaceName(), face.getId());
-                    face.setEmbedding(new Embedding());
-
-                    facesRepository.save(face);
+                    log.error("Error during processing facename {} with id {}", face.getFaceName(), face.getId(), error);
                 }
+
+                face.setEmbedding(faceEmbedding);
+                facesRepository.save(face);
             }
         }
     }
