+++ api/src/main/java/com/exadel/frs/core/trainservice/sdk/faces/feign/dto/ScanResponse.java
@@ -13,18 +13,27 @@
  * or implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package com.exadel.frs.core.trainservice.system.feign.faces.dto;
+
+package com.exadel.frs.core.trainservice.sdk.faces.feign.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.ArrayList;
 import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.experimental.Accessors;
 
 @Data
+@Builder
 @NoArgsConstructor
-public class FindFacesResponse {
+@AllArgsConstructor
+@Accessors(chain = true)
+public class ScanResponse {
 
-    @JsonProperty(value = "plugins_versions")
-    private PluginsVersions pluginsVersions;
-    private List<FindFacesResult> results;
+    @JsonProperty(value = "calculator_version")
+    private String calculatorVersion;
+
+    private List<ScanResult> result = new ArrayList<>();
 }
