+++ api/src/main/java/com/exadel/frs/core/trainservice/service/ScanServiceImpl.java
@@ -20,11 +20,9 @@ import com.exadel.frs.core.trainservice.cache.FaceBO;
 import com.exadel.frs.core.trainservice.cache.FaceCacheProvider;
 import com.exadel.frs.core.trainservice.dao.FaceDao;
 import com.exadel.frs.core.trainservice.entity.Face.Embedding;
-import com.exadel.frs.core.trainservice.exception.NoFacesFoundException;
 import com.exadel.frs.core.trainservice.exception.TooManyFacesException;
-import com.exadel.frs.core.trainservice.system.feign.faces.FacesServiceClient;
-import com.exadel.frs.core.trainservice.system.feign.faces.dto.ScanResponse;
-import feign.FeignException;
+import com.exadel.frs.core.trainservice.sdk.faces.FacesApiClient;
+import com.exadel.frs.core.trainservice.sdk.faces.feign.dto.ScanResponse;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
@@ -38,7 +36,7 @@ public class ScanServiceImpl implements ScanService {
     public static final int MAX_FACES_TO_SAVE = 1;
     public static final int MAX_FACES_TO_RECOGNIZE = 2;
 
-    private final FacesServiceClient facesServiceClient;
+    private final FacesApiClient facesApiClient;
     private final FaceDao faceDao;
     private final FaceCacheProvider faceCacheProvider;
 
@@ -49,12 +47,7 @@ public class ScanServiceImpl implements ScanService {
             final Double detProbThreshold,
             final String modelKey
     ) throws IOException {
-        ScanResponse scanResponse;
-        try {
-            scanResponse = facesServiceClient.scanFaces(file, MAX_FACES_TO_RECOGNIZE, detProbThreshold);
-        } catch (FeignException.BadRequest e) {
-            throw new NoFacesFoundException();
-        }
+        ScanResponse scanResponse = facesApiClient.scanFaces(file, MAX_FACES_TO_RECOGNIZE, detProbThreshold);
         val result = scanResponse.getResult();
 
         if (result.size() > MAX_FACES_TO_SAVE) {
