+++ api/src/test/java/com/exadel/frs/core/trainservice/service/ScanServiceImplTestIT.java
@@ -21,9 +21,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 import com.exadel.frs.core.trainservice.cache.FaceCacheProvider;
 import com.exadel.frs.core.trainservice.dao.FaceDao;
-import com.exadel.frs.core.trainservice.system.feign.faces.FacesServiceClient;
-import com.exadel.frs.core.trainservice.system.feign.faces.dto.ScanResponse;
-import com.exadel.frs.core.trainservice.system.feign.faces.dto.ScanResult;
+import com.exadel.frs.core.trainservice.sdk.faces.FacesApiClient;
+import com.exadel.frs.core.trainservice.sdk.faces.feign.FacesFeignClient;
+import com.exadel.frs.core.trainservice.sdk.faces.feign.dto.ScanResponse;
+import com.exadel.frs.core.trainservice.sdk.faces.feign.dto.ScanResult;
 import com.exadel.frs.core.trainservice.util.MultipartFileData;
 import java.io.IOException;
 import java.util.List;
@@ -36,14 +37,14 @@ import org.springframework.context.annotation.Import;
 import org.springframework.web.multipart.MultipartFile;
 
 @DataJpaTest
-@Import({ScanServiceImpl.class, FacesServiceClient.class, FaceDao.class, FaceCacheProvider.class})
+@Import({ScanServiceImpl.class, FacesFeignClient.class, FaceDao.class, FaceCacheProvider.class})
 class ScanServiceImplTestIT {
 
     @Autowired
     private ScanServiceImpl scanService;
 
     @MockBean
-    private FacesServiceClient facesServiceClient;
+    private FacesApiClient facesFeignClient;
 
     private static final MultipartFile MULTIPART_FILE_DATA = new MultipartFileData("hex-string-1".getBytes(), "test", "application/json");
     private static final String FACE_NAME = "faceName";
@@ -60,7 +61,7 @@ class ScanServiceImplTestIT {
 
     @Test
     public void scanAndFaceTest() throws IOException {
-        when(facesServiceClient.scanFaces(MULTIPART_FILE_DATA, MAX_FACES_TO_RECOGNIZE, THRESHOLD)).thenReturn(SCAN_RESULT);
+        when(facesFeignClient.scanFaces(MULTIPART_FILE_DATA, MAX_FACES_TO_RECOGNIZE, THRESHOLD)).thenReturn(SCAN_RESULT);
 
         val actual = scanService.scanAndSaveFace(MULTIPART_FILE_DATA, FACE_NAME, THRESHOLD, MODEL_KEY);
